added new files for testing and created src/ directory
