+++ src/index.html
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>WebSocket Test</title>
+  </head>
+  <body>
+    <h1>WebSocket Test</h1>
+    <script>
+      const ws = new WebSocket("ws://localhost:3001/ws");
+      ws.onopen = function () {
+        console.log("Received from server");
+        ws.send("wagwan server");
+      };
+      ws.onmessage = function (event) {
+        console.log("Recevied from server:", event.data);
+      };
+      ws.onerror = function (error) {
+        console.log("Websocket error: ", error);
+      };
+      ws.onclose = function () {
+        console.log("WebSocket connection closed");
+      };
+    </script>
+  </body>
+</html>
